webm/pyproject.toml
[tool.poetry]
name = "webm-py"
version = "0.1.0"
description = ""
authors = ["Alexander Demin <ademin@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
av = "^12.1.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
